refactor: move Context
